+++ packages/renderer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLOG
 
+## DEV
+
+1. Features
+    - add `XR` implement
+
 ## 1.7.2
 
 1. Changes